refactor: #1/ exception 패키지 분류
