+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/liked/LikedDuplicatedException.java
@@ -1,4 +1,4 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.liked;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.global.GlobalDuplicatedException;
