+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/liked/LikedNotFoundException.java
@@ -1,11 +1,11 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.liked;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.global.GlobalNotFoundException;
 
-public class ReviewNotFoundException extends GlobalNotFoundException {
+public class LikedNotFoundException extends GlobalNotFoundException {
 
-    public ReviewNotFoundException(ErrorCode errorCode) {
+    public LikedNotFoundException(ErrorCode errorCode) {
         super(errorCode);
     }
 }
