+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/menu/MenuNotFoundException.java
@@ -1,10 +1,10 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.menu;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.global.GlobalNotFoundException;
 
-public class OrderNotFoundException extends GlobalNotFoundException {
-    public OrderNotFoundException(ErrorCode errorCode) {
+public class MenuNotFoundException extends GlobalNotFoundException {
+    public MenuNotFoundException(ErrorCode errorCode) {
         super(errorCode);
     }
 }
