+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/menu/MenuStatusException.java
@@ -1,10 +1,10 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.menu;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.global.GlobalStatusException;
 
-public class StoreStatusException extends GlobalStatusException {
-    public StoreStatusException(ErrorCode errorCode) {
+public class MenuStatusException extends GlobalStatusException {
+    public MenuStatusException(ErrorCode errorCode) {
         super(errorCode);
     }
 }
