+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/order/OrderStatusException.java
@@ -1,10 +1,10 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.order;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.global.GlobalStatusException;
 
-public class ReviewStatusException extends GlobalStatusException {
-    public ReviewStatusException(ErrorCode errorCode) {
+public class OrderStatusException extends GlobalStatusException {
+    public OrderStatusException(ErrorCode errorCode) {
         super(errorCode);
     }
 }
