+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/store/StoreDuplicatedException.java
@@ -1,4 +1,4 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.store;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.global.GlobalDuplicatedException;
