+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/store/StoreStatusException.java
@@ -1,10 +1,10 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.store;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.global.GlobalStatusException;
 
-public class OrderStatusException extends GlobalStatusException {
-    public OrderStatusException(ErrorCode errorCode) {
+public class StoreStatusException extends GlobalStatusException {
+    public StoreStatusException(ErrorCode errorCode) {
         super(errorCode);
     }
 }
