+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/user/UserDuplicatedException.java
@@ -1,12 +1,11 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.user;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.global.GlobalDuplicatedException;
 
-public class UserNotExistException extends GlobalDuplicatedException {
+public class UserDuplicatedException extends GlobalDuplicatedException {
 
-    public UserNotExistException(ErrorCode errorCode) {
+    public UserDuplicatedException(ErrorCode errorCode) {
         super(errorCode);
     }
-
 }
