+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/user/UserNotExistException.java
@@ -1,11 +1,12 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.user;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.global.GlobalDuplicatedException;
 
-public class UserDuplicatedException extends GlobalDuplicatedException {
+public class UserNotExistException extends GlobalDuplicatedException {
 
-    public UserDuplicatedException(ErrorCode errorCode) {
+    public UserNotExistException(ErrorCode errorCode) {
         super(errorCode);
     }
+
 }
