+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/user/UserPasswordMismatchException.java
@@ -1,4 +1,4 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.user;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.global.GlobalMismatchException;
