+++ src/main/java/com/sparta/delivery_app/common/globalcustomexception/user/UserStatusException.java
@@ -1,10 +1,10 @@
-package com.sparta.delivery_app.common.globalcustomexception;
+package com.sparta.delivery_app.common.globalcustomexception.user;
 
 import com.sparta.delivery_app.common.exception.errorcode.ErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.global.GlobalStatusException;
 
-public class MenuStatusException extends GlobalStatusException {
-    public MenuStatusException(ErrorCode errorCode) {
+public class UserStatusException extends GlobalStatusException {
+    public UserStatusException(ErrorCode errorCode) {
         super(errorCode);
     }
 }
