+++ src/main/java/com/sparta/delivery_app/domain/liked/adapter/LikedAdapter.java
@@ -1,7 +1,7 @@
 package com.sparta.delivery_app.domain.liked.adapter;
 
 import com.sparta.delivery_app.common.exception.errorcode.LikedErrorCode;
-import com.sparta.delivery_app.common.globalcustomexception.LikedNotFoundException;
+import com.sparta.delivery_app.common.globalcustomexception.liked.LikedNotFoundException;
 import com.sparta.delivery_app.domain.liked.entity.Liked;
 import com.sparta.delivery_app.domain.liked.repository.LikedRepository;
 import com.sparta.delivery_app.domain.store.entity.Store;
