+++ src/main/java/com/sparta/delivery_app/domain/menu/adapter/MenuAdapter.java
@@ -1,7 +1,7 @@
 package com.sparta.delivery_app.domain.menu.adapter;
 
 import com.sparta.delivery_app.common.exception.errorcode.MenuErrorCode;
-import com.sparta.delivery_app.common.globalcustomexception.MenuNotFoundException;
+import com.sparta.delivery_app.common.globalcustomexception.menu.MenuNotFoundException;
 import com.sparta.delivery_app.domain.menu.entity.Menu;
 import com.sparta.delivery_app.domain.menu.entity.MenuStatus;
 import com.sparta.delivery_app.domain.menu.repository.MenuRepository;
