+++ src/main/java/com/sparta/delivery_app/domain/store/adapter/StoreAdapter.java
@@ -1,9 +1,9 @@
 package com.sparta.delivery_app.domain.store.adapter;
 
 import com.sparta.delivery_app.common.exception.errorcode.StoreErrorCode;
-import com.sparta.delivery_app.common.globalcustomexception.StoreDuplicatedException;
-import com.sparta.delivery_app.common.globalcustomexception.StoreNotFoundException;
-import com.sparta.delivery_app.common.globalcustomexception.StoreRegisteredHistoryException;
+import com.sparta.delivery_app.common.globalcustomexception.store.StoreDuplicatedException;
+import com.sparta.delivery_app.common.globalcustomexception.store.StoreNotFoundException;
+import com.sparta.delivery_app.common.globalcustomexception.store.StoreRegisteredHistoryException;
 import com.sparta.delivery_app.domain.store.dto.request.RegisterStoreRequestDto;
 import com.sparta.delivery_app.domain.store.entity.Store;
 import com.sparta.delivery_app.domain.store.repository.StoreRepository;
